Update mistral-7b.en.mdx
"for faster inference and reduce..." changed to "for faster inference and reduced...". (added 'd')

"we escaped the code segments by to display" changed to "we escaped the code segments to display". (removed 'by')

"Let's start with a simple example and instruct the model to achieve a simple task based on an instruction." changed to "Let's start with a simple example and instruct the model to achieve a task based on an instruction." (removed second 'simple' for readability)

"The model is prone to common prompt injections, here are some examples:" Changed to "The model is prone to common prompt injections; here are some examples:" (changed a colon to a semicolon)